// src/types/IncrementorDecrementor.d.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import { ContractOptions } from "web3-eth-contract";
import { EventLog } from "web3-core";
import { EventEmitter } from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface IncrementorDecrementor extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): IncrementorDecrementor;
  clone(): IncrementorDecrementor;
  methods: {
    increment1(): PayableTransactionObject<void>;

    increment10(): PayableTransactionObject<void>;

    increment100(): PayableTransactionObject<void>;

    increment1000(): PayableTransactionObject<void>;

    decrement1(): PayableTransactionObject<void>;

    decrement10(): PayableTransactionObject<void>;

    decrement100(): PayableTransactionObject<void>;

    decrement1000(): PayableTransactionObject<void>;

    result(): NonPayableTransactionObject<string>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}